python/google_slides_quickstart.py
import os.path
from googleapiclient.discovery import build
from googleapiclient.errors import HttpError
from google.oauth2 import service_account

SCOPES = ['https://www.googleapis.com/auth/presentations']
SERVICE_ACCOUNT_FILE = 'credentials.json'

#Authenticate to the slides after 
creds = None
creds = service_account.Credentials.from_service_account_file(
        SERVICE_ACCOUNT_FILE, scopes=SCOPES)


# The ID of the presentation to edit
PRESENTATION_ID = '1RPl2YlqPdWKIfWvt-Vc3NwseYNq6SsGslSeR56-Dfuk'


# If there are no (valid) credentials available, let the user log in.

try:
    service = build('slides', 'v1', credentials=creds)

    # Call the Slides API
    presentation = service.presentations().get(
        presentationId=PRESENTATION_ID).execute()
    slides = presentation.get('slides')

    print('The presentation contains {} slides:'.format(len(slides)))
    for i, slide in enumerate(slides):
        print('- Slide #{} contains {} elements.'.format(
            i + 1, len(slide.get('pageElements'))))
except HttpError as err:
    print(err)


