requirements.txt
google_api_python_client==2.39.0
google_auth_oauthlib==0.5.0
protobuf==3.19.4
requests==2.18.4
